+++ src/main.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './app/app';
-import '@/app/services/dayjs';
+import '@/shared/dayjs';
 
 import '@/app/styles/index.scss';
