+++ src/widgets/course-main/course-main.test.tsx
@@ -1,9 +1,9 @@
 import { screen } from '@testing-library/react';
 import { CourseMain } from './course-main';
-import { dayJS } from '@/app/services/dayjs';
 import { CourseStatus } from '@/app/types';
 import { MOCKED_IMAGE_PATH } from '@/shared/__tests__/constants';
 import { renderWithRouter } from '@/shared/__tests__/utils';
+import { dayJS } from '@/shared/dayjs';
 
 vi.mock('@/app/hooks/use-course-by-title');
 vi.mock('react-router-dom', async (importOriginal) => {
